requirements.txt
Django==1.4.2
Pillow
South==0.7.6
Werkzeug
django-classy-tags==0.3.4.1
django-debug-toolbar
django-extensions
django-filebrowser
django-mptt==0.5.4
django-polymorphic==0.2
django-shop==0.1.1
django-shop-categories==0.8.beta.2
django-sitetree==0.9.1
django-tinymce==1.5.1b2
django-treeadmin==0.2.1
django-classbasedsettings
django-singletons
django-shop-configurableproduct
ipython
sorl-thumbnail==11.12
wsgiref
